Refactor person.rb file and create new class for each file and collect all classes to src folder
