--- person.rb
@@ -1,65 +1,0 @@
-require './rental'
-
-class Nameable
-  def correct_name
-    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
-  end
-end
-
-class Decorator < Nameable
-  attr_accessor :nameable
-
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    name = @nameable.correct_name
-    name.length > 10 ? name[0..9] : name
-  end
-end
-
-class Person < Nameable
-  attr_reader :id
-  attr_accessor :name, :age, :rentals
-
-  def initialize(age, name = 'Unknown', parent_permission: true)
-    super()
-    @id = rand(1..1000)
-    @name = name
-    @age = age
-    @parent_permission = parent_permission
-    @rentals = []
-  end
-
-  def can_use_services?
-    of_age? || @parent_permission
-  end
-
-  def correct_name
-    @name
-  end
-
-  def rent(date, book)
-    Rental.new(date, book, self)
-  end
-
-  private
-
-  def of_age?
-    @age >= 18
-  end
-end
