+++ src/book.rb
@@ -1,4 +1,4 @@
-require './rental'
+require './src/rental'
 
 class Book
   attr_accessor :title, :author, :rentals
