+++ src/capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require './src/decorator'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
