+++ src/trimmer_decorator.rb
@@ -1,0 +1,8 @@
+require './src/decorator'
+
+class TrimmerDecorator < Decorator
+  def correct_name
+    name = @nameable.correct_name
+    name.length > 10 ? name[0..9] : name
+  end
+end
